+++ info.json
@@ -1,6 +1,6 @@
 {
 	"name": "hotpatch-multimod",
-	"version": "1.1.18",
+	"version": "1.1.19",
 	"title": "Server-side Multi-mod Scenario",
 	"author": "Chrisgbk",
 	"homepage": "",
