+++ scenarios/Hotpatch/locale/ar/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=.بكوكلا اذه ناكس م كسفن يمحت نا ىسنت الو ,ةيكيتاموتوألا ربع كقيرط ينبتسو ,رفصلا نم أدبتس .كلت خيراوصلا قالطإ ةطحم حتف عيطتست فيكل ةروطتملا ايجولونكتلل ثاحبأب مايقلا ىلإ جاتحتس .اهنم خوراص قالطإ مث خيراوص قالطإ ةطحم ءانب قيرط نع كلذب مايقلا كنكمي ,ءاضفلا ىلإ اخوراص قلطت نا يه كتمهم ,factorioـل رحلا بعللا روط وه اذه
+msg-intro=.بكوكلا اذه ناكس نم كسفن يمحت نا ىسنت الو ,ةممتألا ربع كقيرط ينبتسو ,رفصلا نم أدبتس .خيراوصلا ةصنم حتف عيطتست فيكل ةمدقتم ةينقت ثاحبأب مايقلا ىلإ جاتحتس .ءاضفلا ىلإ اخوراص قلطت نأ يه كتمهم .factorioـل رحلا بعللا روط وه اذه
 scenario-name=رحلا بعللا
