+++ scenarios/Hotpatch/locale/bg/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Това е режимът на свободна игра. Задачата ви е да изстреляте ракета в космоса. За това ще е нужно да построите ракетна шахта и изстреляте ракетата със спътник. Ще трябва да проучите развити технологии, които ще отключат ракетната шахта. Започнете с малко и постепенно автоматизирайте процеса, като не забравяте да се защитавате от местните форми на живот.
+msg-intro=Това е режимът на свободна игра. Задачата ви е да изстреляте ракета в космоса. Ще трябва да проучите напреднали технологии, които ще отключат ракетната шахта. Започнете с малко и постепенно автоматизирайте процеса, като не забравяте да се защитавате от местните форми на живот.
 scenario-name=Свободна игра
