+++ scenarios/Hotpatch/locale/cs/freeplay.cfg
@@ -1,3 +1,2 @@
-msg-intro=Toto je volná hra Factorio. Tvým úkolem je vypustit raketu do vesmíru. Postav raketové silo a vypusť raketu. Pro odemknutí raketového sila budeš muset vyzkoumat pokročilé technologie. Začni v malém, postupuj s automatizací a nezapomeň se chránit před domorodci.
 scenario-name=Volná hra
