+++ scenarios/Hotpatch/locale/da/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Dette er Factorio freeplay. Din opgave er at affyre en raket ud i rummet. Gør dette ved at konstruere en raketsilo og affyr en raket. Du bliver nød til at undersøge avancerede teknologier for at låse op for raketsiloen. Start småt og arbejd din vej op med automatisering og glem ikke at beskytte dig selv fra de indfødte.
+msg-intro=Dette er Factorio freeplay. Din opgave er at sende en raket ud i rummet. Gør dette ved at konstruere en raketsilo og affyr en raket. Du bliver nød til at undersøge avancerede teknologier for at låse op for raketsiloen. Start småt, og arbejd din vej op med automatisering og glem ikke at beskytte dig selv fra de indfødte.
 scenario-name=Fri lej
