+++ scenarios/Hotpatch/locale/de/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Du befindest dich im freien Spiel Modus von Factorio. Ziel ist es, eine Rakete ins Weltall zu schicken. Baue hierzu das Raketensilo um die Rakete starten zu können. Um das Raketensilo freischalten zu können wirst du fortgeschrittene Technologien erforschen müssen. Fange klein an und erforsche die Automation. Vergiss aber nicht, dich vor den Ureinwohner zu schützen.
+msg-intro=Dies ist das Freie Spiel in Factorio. Ziel ist es, eine Rakete in den Weltraum zu senden. Du musst fortschrittliche Technologien erforschen, um den Raketensilo freizuschalten. Fange klein an, lerne die Automatisierung kennen und vergiss nicht, dich vor den Eingeborenen zu schützen.
 scenario-name=Freies Spiel
