+++ scenarios/Hotpatch/locale/en/freeplay.cfg
@@ -1,2 +1,2 @@
-msg-intro=This is the Factorio freeplay. Your task is to launch a rocket into space. Do this by constructing a rocket silo and launching a rocket. You will need to research advanced technologies in order to unlock the rocket silo. Start small, work your way up with automation and don't forget to protect yourself from the natives.
+msg-intro=This is the Factorio freeplay. Your task is to launch a rocket into space. You will need to research advanced technologies in order to unlock the rocket silo. Start small, work your way up with automation and don't forget to protect yourself from the natives.
 scenario-name=Freeplay
