+++ scenarios/Hotpatch/locale/en/hotpatch.cfg
@@ -1,6 +1,6 @@
 [hotpatch]
 log=@__1__:__2__] __3__
-log-mod=@__1__:__2__][__4__] __3__
+log-mod=@__1__:__2__ >__4__<] __3__
 severe=SEVERE ERROR: __1__
 error=ERROR: __1__
 warning=WARNING: __1__
@@ -37,6 +37,9 @@ event-registering=registering events...
 on-tick-event-registered=registered on_tick event...
 on-event-registered=registered event __1__...
 on-nth-tick-event-registered=registered nth_tick event __1__...
+on-tick-event-unregistered=unregistered on_tick event...
+on-event-unregistered=unregistered event __1__...
+on-nth-tick-event-unregistered=unregistered nth_tick event __1__...
 event-running=running event __1__...
 nth-tick-event-running=running nth_event __1__...
 mod-on-init=running on_init...
