+++ scenarios/Hotpatch/locale/es-ES/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Este es el modo de juego libre de Factorio. Tu tarea es la de lanzar un cohete al espacio. Logra esto construyendo un Silo de Cohetes y lanzando un cohete con un satélite. Necesitarás investigar tecnologías avanzadas para poder desbloquear el Silo de Cohetes. Comienza poco a poco, trabaja en tu sistema de automatización y no te olvides de protegerte de los nativos.
+msg-intro=Este es el Modo libre de Factorio. Tu misión es lanzar un cohete al espacio. Para lograr tu objetivo, necesitarás investigar tecnologías avanzadas. Desde el inicio usa la automatización para avanzar gradualmente hacia el lanzamiento y no se olvides de protegerse de los nativos.
 scenario-name=Modo libre
