+++ scenarios/Hotpatch/locale/et/freeplay.cfg
@@ -1,2 +1,3 @@
+msg-intro=See on Factorio vabamäng. Sinu eesmärgiks on saata rakett avakosmosesse. Seda saad teha ehitades raketi silotorni ja lennutades raketi koos satelliidiga. Selleks pead õppima selgeks keerukad tehnoloogiaharud, et raketi silotorni lahti lukustada. Alusta väikselt, tööta end üles automatiseerimise abil ja ära unusta ennast kaitsta kohalike eluvormide eest.
 scenario-name=Vabamäng
