+++ scenarios/Hotpatch/locale/ga-IE/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Fáilte go dtí an saor-imeartha Factorio. Tá tú mar aidhm agat chun roicéid a lainseáil go spás. Is féidir leat é seo a dhéanamh trí thógáil Sadhlann Roicéid agus le satailít a chur leis an roicéad roimh á lainseáil. Teastaíonn taighde déanta ar ardteicneolaíochtaí agat chun na Sadhlainne Roicéid a dhíghlasáil. Tosaigh le bunáit bheag, méadú agus oibriú chun bairr tríd an uathoibríocht, agus ná déan dearmad chun túsa féin agus do bhunáit a chosaint i gcoinne na ndúchasach.
+msg-intro=Fáilte go dtí an saor-imeartha Factorio. Tá tú mar aidhm agat chun roicéid a lainseáil go spás. Teastaíonn taighde déanta ar ardteicneolaíochtaí agat chun na Sadhlainne Roicéid a dhíghlasáil. Tosaigh le bunáit bheag, méadú agus oibriú chun bairr tríd an uathoibríocht, agus ná déan dearmad chun túsa féin agus do bhunáit a chosaint i gcoinne na ndúchasach.
 scenario-name=Saor-imirt
