+++ scenarios/Hotpatch/locale/id/freeplay.cfg
@@ -1,0 +1,3 @@
+msg-intro=Ini adalah Factorio freeplay. Tugasmu adalah meluncurkan roket ke angkasa luar. Kamu membutuhkan penelitian canggih berteknologi agar tidak mengunci roket silo. Mulai dari yang kecil, tingkatkan otomatisasi dan jangan lupa melindungi dirimu dari native.
+scenario-name=Freeplay
+
