+++ scenarios/Hotpatch/locale/pl/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Oto swobodna rozgrywka Factorio. Twoim zadaniem jest wysłanie rakiety w kosmos. Dokonaj tego przez skonstruowanie silosu rakietowego i wystrzelenie rakiety. Żeby odblokować silos rakietowy będziesz musiał odkryć zaawansowane technologie. Rozpocznij od podstaw, rozwijaj się wykorzystując automatyzację i nie zapomnij o obronie przed rodzimymi mieszkańcami planety.
+msg-intro=To jest gra swobodna Factorio. Twoim zadaniem jest wysłanie rakiety w kosmos. Będziesz musiał zbadać zaawansowane technologie, aby odblokować silos rakietowy. Zacznij od czegoś małego, rozwijaj się używając automatyzacji i nie zapomnij o obronie przed tubylcami.
 scenario-name=Gra swobodna
