+++ scenarios/Hotpatch/locale/pt-BR/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Este é o modo livre de Factorio. Sua tarefa é lançar um foguete no espaço. Faça isso construindo um silo de foguetes e lançando um foguete. Você precisará pesquisar tecnologias avançadas para desbloquear o silo de foguetes. Comece pequeno, trabalhe seu caminho com automação e não se esqueça de se proteger dos nativos.
+msg-intro=Este é o modo livre de Factorio. Sua tarefa é lançar um foguete no espaço. Você deve pesquisar tecnologias avançadas para desbloquear o Centro de Lançamento. Comece do zero, trace seu caminho através da automação e não se esqueça de se proteger contra os nativos.
 scenario-name=Jogo livre
