+++ scenarios/Hotpatch/locale/pt-PT/freeplay.cfg
@@ -1,3 +1,2 @@
-msg-intro=Este é o modo livre do Factorio. O teu objetivo é lançar um foguetão para o espaço. Faz isto, contruindo um silo de foguetão e lançando-o.\nVais precisar de pesquisar tecnologias avançadas para desploquear o silo de foguetão. Começa simples, e trabalha e evolui com automação mas não te esqueças de proteger do nativos.
 scenario-name=Modo Livre
