+++ scenarios/Hotpatch/locale/sk/freeplay.cfg
@@ -1,3 +1,2 @@
-msg-intro=Toto je voľná hra Factorio. Tvojou úlohou je vypustiť raketu do vesmíru. Postav raketové silo a vypusti raketu. Pre odomknutie raketového sila budeš musieť vyskúmať pokročilé technológie. Začni v malom, postupuj s automatizáciou a nezabnudni sa chrániť pred domácimi.
 scenario-name=Voľná hra
