+++ scenarios/Hotpatch/locale/sv-SE/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Det här är Factorio frispel. Ditt uppdrag är att avfyra en raket ut i rymden. Gör det genom att tillverka en raket silo och avfyra en raket med en satellit. Du kommer att behöva forska fram avancerade teknologier för att kunna låsa upp raket silon. Börja smått, jobba dig uppåt med hjälp av automatisering och glöm inte att försvara dig själv mot infödingarna.
+msg-intro=Det här är Factorio frispel. Ditt uppdrag är att avfyra en raket ut i rymden. Du kommer att behöva forska fram avancerade teknologier för att kunna låsa upp raket silon. Börja smått, jobba dig uppåt med hjälp av automatisering och glöm inte att försvara dig själv mot infödingarna.
 scenario-name=Frispel
