+++ scenarios/Hotpatch/locale/tr/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=Bu Factorio'nun serbest oynanışıdır. Göreviniz uzaya bir roket göndermek. Bunu roket rampası inşa edip ve Roket fırlatarak yapınız. Roket rampasını açmak için gelişmiş teknolojileri araştırmanız gerekecek. Ufaktan başlayın ve otomasyonla büyüyün, ve yerlilerden kendiniz korumayı unutmayın.
-scenario-name=Serbest Oyun
+msg-intro=Bu Factorio'nun serbest oynanışıdır. göreviniz uzaya bir roket göndermek. Rocket rampasını açmak için gelişmiş teknolojileri araştırmanız gerekecek. Ufaktan başlayın ve otomasyonla büyüyün, ve yerlilerden kendinizi korumayı unutmayın.
+scenario-name=Serbest oyun
