+++ scenarios/Hotpatch/locale/zh-CN/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=欢迎来到异星工厂标准模式，你的目标是要把火箭发射到太空。为此你需要研发一系列高等科技，修建火箭发射井，最后发射火箭。从小作坊生产起步，逐渐建立起属于自己的自动化大工厂吧！同时，别忘了防范异星虫子们的骚扰 ...
+msg-intro=欢迎来到异星工厂标准模式。你的目标是把火箭射入太空，为此你要以火箭发射井为方向研发一系列高等科技。从小作坊生产起步，逐渐建立起属于自己的自动化大工厂吧！同时，别忘了防范异星虫子的骚扰 ...
 scenario-name=标准模式
