+++ scenarios/Hotpatch/locale/zh-TW/freeplay.cfg
@@ -1,3 +1,3 @@
-msg-intro=這是異星工廠的自由模式。你的任務是發射火箭探索外太空。請建造火箭發射井並發射火箭。你會需要研究大量的進階科技以解鎖火箭發射井。從頭開始，靠著自動化的協助邁向目標的同時，也別忘了從外星住民手中保護你自己。
+msg-intro=這是異星工廠的自由模式，你的任務是發射火箭上太空。你將需要研究高等科技，並解鎖火箭發射井。從頭開始，利用自動化來逐步邁向目標，但也別忘記防止被原生種襲擊。
 scenario-name=自由模式
